code/requirements.txt
opencv_python==4.5.2.52
numpy==1.19.4
matplotlib==3.3.3
torch==1.7.0+cu101
torchvision==0.8.1+cu101
pandas==1.1.5
tqdm==4.61.0
Pillow==8.3.1
tensorboardX==2.4
